requirements.txt
# Microsoft Graph SDK and Azure Identity
msgraph-sdk
azure-identity

# RabbitMQ client
pika

# For loading .env files
python-dotenv

# For async operations
asyncio